Crate stahl_util

Some utilities used throughout the Stahl compiler.

Structs

SharedPath
A path whose backing storage is shared (via a reference count).

SharedString
A string whose backing storage is shared (via a reference count).

SplitVec
A vector which is split at a given index. Effectively, this is a Vec with a single uninitialized position, which can be reunified to yield the original Vec without memory allocation.

Taker
A wrapper for values which can be taken out.

Functions

display_vec
Returns a Display for printing a Vec.

fmt_char
Formats a string character as per Stahl's escaping rules.

fmt_iter
Formats a list of space-separated items.

fmt_string
Formats a string as per Stahl's escaping rules.

genint
Generates a unique integer.

unwrap_arc
Takes a value from an Arc if the reference is unique, otherwise clones it out.

unwrap_rc
Takes a value from an Rc if the reference is unique, otherwise clones it out.

unzip_result_iter
Unzips an iterator that returns its pairs inside a Result.